+++ packages/schema-definition/src/index.ts
@@ -1,6 +1,6 @@
 import { AllowAllPermissionFactory } from '@contember/schema-utils'
 export * from './model'
-import PermissionsBuilder from './acl/PermissionsBuilder'
+import PermissionsBuilder from './acl/builder/PermissionsBuilder'
 import * as InputValidation from './validation'
 
 export { AllowAllPermissionFactory, PermissionsBuilder, InputValidation }
